fix git clone url
Thanks for this great plugin! I founda minor error in the installation instructions. It's an easy spot for experienced coders, I'm sure, but I'm submitting a change request in case it helps neophytes like me. Thanks again for sharing this plugin! It's really helped with some of our organizational workflows.